requirements.txt
Django==5.2.6
PyJWT